LayoutTests/http/tests/workers/service/tainted-image-fetch-expected.txt


Registering service worker
Service worker registered
Status is Got response for http://127.0.0.1:8000/workers/service/resources/cors-image-fetch-iframe.html, status code is 200
Loading image
PASS: Loaded image
Status is Got response for http://localhost:8000/resources/square100.png, status code is 200
Image size: 100x100
PASS: Image is not tainted
